rename packages, cleanup
