+++ BackEnd/core/src/main/java/GroceryFamily/GroceryElders/Config.java
@@ -1,0 +1,12 @@
+package GroceryFamily.GroceryElders;
+
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@Configuration
+@ComponentScan(basePackages = {"GroceryFamily"})
+@EnableJpaRepositories(basePackages = {"GroceryFamily"})
+@EntityScan(basePackages = {"GroceryFamily"})
+public class Config {}
