+++ BackEnd/scraper/pom.xml
@@ -10,7 +10,7 @@
         <version>0.1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>FinalProject-scraper</artifactId>
+    <artifactId>GroceryDad</artifactId>
 
     <profiles>
         <profile>
@@ -49,7 +49,13 @@
     <dependencies>
         <dependency>
             <groupId>GroceryFamily</groupId>
-            <artifactId>FinalProject-core</artifactId>
+            <artifactId>GroceryElders</artifactId>
+        </dependency>
+
+        <!-- https://github.com/spring-projects -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-jpa</artifactId>
         </dependency>
 
         <!-- https://github.com/SeleniumHQ/selenium -->
@@ -76,11 +82,7 @@
             <artifactId>jackson-databind</artifactId>
         </dependency>
 
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-jpa</artifactId>
-        </dependency>
-
+        <!-- https://github.com/spring-projects :: test -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
