+++ BackEnd/scraper/src/main/java/GroceryFamily/GroceryDad/Config.java
@@ -1,0 +1,23 @@
+package GroceryFamily.GroceryDad;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+import java.time.Duration;
+
+@Configuration
+@Import(value = {GroceryFamily.GroceryElders.Config.class})
+public class Config {
+    @Bean
+    WebDriver webDriver() {
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--remote-allow-origins=*");
+        WebDriver driver = new ChromeDriver(options);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+        return driver;
+    }
+}
