+++ BackEnd/scraper/src/main/java/GroceryFamily/GroceryDad/Parser/PrismaParser.java
@@ -1,0 +1,7 @@
+package GroceryFamily.GroceryDad.Parser;
+
+public class PrismaParser{
+    private final String prismaUrl= "prismaperemarket.ee";
+
+
+}
