+++ BackEnd/scraper/src/main/java/GroceryFamily/GroceryDad/Parser/RimiParser.java
@@ -1,0 +1,10 @@
+package GroceryFamily.GroceryDad.Parser;
+
+public class RimiParser {
+    private final String rimiUrl = "rimi.ee";
+
+
+//    public RimiParser(WebDriver driver, GroceryInfoService groceryInfoService) {
+//        super(driver, groceryInfoService);
+//    }
+}
