+++ BackEnd/scraper/src/main/java/GroceryFamily/GroceryDad/productCheckOnThePage/NameChecking.java
@@ -1,0 +1,20 @@
+package GroceryFamily.GroceryDad.productCheckOnThePage;
+
+import GroceryFamily.GroceryElders.model.Product;
+
+import java.util.List;
+
+
+public class NameChecking {
+
+
+
+    //check what names of product we have on the page, if in collection list of Products from the page don`t contains
+    // similar name like in DB, then remove
+    public static void checkingName(List<Product> onThePage, List<String> groceriesName) {
+
+    }
+
+
+
+}
