+++ BackEnd/scraper/src/main/java/GroceryFamily/GroceryDad/productCheckOnThePage/PriceChecking.java
@@ -1,0 +1,4 @@
+package GroceryFamily.GroceryDad.productCheckOnThePage;
+
+public class PriceChecking {
+}
