+++ BackEnd/scraper/src/main/java/GroceryFamily/GroceryDad/service/WebScraperServiceDubl.java
@@ -1,0 +1,32 @@
+package GroceryFamily.GroceryDad.service;
+//
+//
+//import com.example.FinalProject.Parser.BarboraParser;
+//import com.example.FinalProject.Parser.PrismaParser;
+//import com.example.FinalProject.Parser.RimiParser;
+//import com.example.FinalProject.model.Product;
+//
+//import java.util.List;
+//
+public class WebScraperServiceDubl {
+//
+//    // From parse package
+//   private RimiParser rimiParser;
+//   private BarboraParser barboraParser;
+//   private PrismaParser prismaParser;
+////    BarboraParser barboraParser = new BarboraParser();
+//
+////    List<Product> rimiProduct = rimiParser.parserProducts();
+////    List<Product> prismaProduct = prismaParser.parserProducts();
+////    List<Product> barboraProduct = barboraParser.parserProducts();
+//
+//    //another actions
+//    private void performAdditionalOperations() {
+//
+//    }
+//
+//    private void displayResults(){
+//
+//    }
+//
+}
