--- BackEnd/scraper/src/main/java/com/example/FinalProject/Parser/PrismaParser.java
@@ -1,10 +1,0 @@
-package com.example.FinalProject.Parser;
-
-
-
-
-public class PrismaParser{
-    private final String prismaUrl= "prismaperemarket.ee";
-
-
-}
