--- BackEnd/scraper/src/main/java/com/example/FinalProject/Parser/RimiParser.java
@@ -1,14 +1,0 @@
-package com.example.FinalProject.Parser;
-
-import com.example.FinalProject.service.GroceryInfoService;
-import org.openqa.selenium.WebDriver;
-
-
-public class RimiParser {
-    private final String rimiUrl = "rimi.ee";
-
-
-//    public RimiParser(WebDriver driver, GroceryInfoService groceryInfoService) {
-//        super(driver, groceryInfoService);
-//    }
-}
