--- BackEnd/scraper/src/main/java/com/example/FinalProject/config/ApplicationConfiguration.java
@@ -1,22 +1,0 @@
-package com.example.FinalProject.config;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.time.Duration;
-
-@Configuration
-public class ApplicationConfiguration {
-    @Bean
-    WebDriver chromeDriver() {
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--remote-allow-origins=*");
-        WebDriver driver = new ChromeDriver(options);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
-        return driver;
-
-    }
-}
