--- BackEnd/scraper/src/main/java/com/example/FinalProject/productCheckOnThePage/NameChecking.java
@@ -1,20 +1,0 @@
-package com.example.FinalProject.productCheckOnThePage;
-
-import com.example.FinalProject.model.Product;
-import com.example.FinalProject.similarity.ProductFilter;
-import java.util.List;
-
-
-public class NameChecking {
-
-
-
-    //check what names of product we have on the page, if in collection list of Products from the page don`t contains
-    // similar name like in DB, then remove
-    public static void checkingName(List<Product> onThePage, List<String> groceriesName) {
-
-    }
-
-
-
-}
