--- BackEnd/scraper/src/main/java/com/example/FinalProject/productCheckOnThePage/PriceChecking.java
@@ -1,4 +1,0 @@
-package com.example.FinalProject.productCheckOnThePage;
-
-public class PriceChecking {
-}
