+++ BackEnd/scraper/src/test/java/GroceryFamily/GroceryDad/ScraperTest.java
@@ -1,0 +1,12 @@
+package GroceryFamily.GroceryDad;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = Scraper.class)
+class ScraperTest {
+    @Test
+    void test() {
+        // todo: implement some better test
+    }
+}
