+++ BackEnd/server/pom.xml
@@ -10,7 +10,7 @@
         <version>0.1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>FinalProject-server</artifactId>
+    <artifactId>GroceryMom</artifactId>
 
     <profiles>
         <profile>
@@ -49,9 +49,10 @@
     <dependencies>
         <dependency>
             <groupId>GroceryFamily</groupId>
-            <artifactId>FinalProject-core</artifactId>
+            <artifactId>GroceryElders</artifactId>
         </dependency>
 
+        <!-- https://github.com/spring-projects -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-data-jpa</artifactId>
@@ -61,6 +62,7 @@
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
 
+        <!-- https://github.com/spring-projects :: test -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
