+++ BackEnd/server/src/main/java/GroceryFamily/GroceryMom/Config.java
@@ -1,0 +1,8 @@
+package GroceryFamily.GroceryMom;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import(value = {GroceryFamily.GroceryElders.Config.class})
+public class Config {}
