+++ BackEnd/server/src/test/java/GroceryFamily/GroceryMom/ServerTest.java
@@ -1,0 +1,12 @@
+package GroceryFamily.GroceryMom;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = Server.class)
+class ServerTest {
+    @Test
+    void test() {
+        // todo: implement some better test
+    }
+}
