--- BackEnd/server/src/test/java/com/example/FinalProject/FinalProjectApplicationTests.java
@@ -1,12 +1,0 @@
-package com.example.FinalProject;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = FinalProjectApplication.class)
-class FinalProjectApplicationTests {
-    @Test
-    void test() {
-        // todo: implement some better test
-    }
-}
